Manually play notification sound on Marshmallow or earlier
On Marshmallow and earlier, notifications aren't playing any sound.
There's no problem on Lollipop or later, but I've been unable to track
down the root cause.

This workaround fixes the problem, but isn't very satisfying.

The READ_EXTERNAL_STORAGE permission is required for RingtoneManager
to fetch the ringtone before we play it.  Users have to manually grant
this permission for it to work.  Otherwise, the system crashes when it
tries to play the sound.